create git package
